--- app/src/main/java/com/alvindizon/panahon/di/RepositoryModule.kt
@@ -1,21 +1,0 @@
-package com.alvindizon.panahon.di
-
-import com.alvindizon.panahon.home.integration.HomeViewRepository
-import com.alvindizon.panahon.integration.HomeViewRepositoryImpl
-import com.alvindizon.panahon.repo.PanahonRepo
-import com.alvindizon.panahon.repo.PanahonRepoImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class RepositoryModule {
-
-    @Binds
-    abstract fun bindPanahonRepo(repo: PanahonRepoImpl): PanahonRepo
-
-    @Binds
-    abstract fun bindHomeViewRepo(repo: HomeViewRepositoryImpl): HomeViewRepository
-}
