+++ feature/home/src/main/java/com/alvindizon/panahon/home/usecase/CheckLocationIsOnUseCase.kt
@@ -2,7 +2,6 @@ package com.alvindizon.panahon.home.usecase
 
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.IntentSenderRequest
-import com.alvindizon.panahon.home.integration.HomeViewRepository
 import com.alvindizon.panahon.location.LocationManager
 import javax.inject.Inject
 import javax.inject.Singleton
